+++ docs/changelog.md
@@ -7,6 +7,8 @@ permalink: /changelog/
 # 145.0 (In Development)
 * **support-ktx**
   * 🚒 Bug fixed [Bug 1979064](https://bugzilla.mozilla.org/show_bug.cgi?id=1979064). Added back the API checks for `Window.setupPersistentInsets` and `ImeInsetsSynchronizer.setup` as on Android <13 the insets framework is not reliable.
+* **support-appservices**
+  * ⚠️ **Breaking change**: Updated the `AppServicesInitializer.init` to take in a configuration object instead of individual components.
 
 # 144.0
 * **feature-customtabs**